Added JWT functionality

+++ src/main/kotlin/quantsage/Spring_boot_backend/controllers/NoteController.kt
@@ -1,7 +1,7 @@
 package quantsage.Spring_boot_backend.controllers
 
 import org.bson.types.ObjectId
-import org.springframework.data.annotation.Id
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.*
 import quantsage.Spring_boot_backend.database.model.Note
 import quantsage.Spring_boot_backend.database.repository.NoteRepository
@@ -10,7 +10,8 @@ import java.time.Instant
 @RestController
 @RequestMapping("/notes")
 class NoteController(
-    private val repository: NoteRepository
+    private val repository: NoteRepository,
+    private val noteRepository: NoteRepository
 ) {
 
     data class NoteRequest(
@@ -32,6 +33,7 @@ class NoteController(
     fun save(
         @RequestBody body : NoteRequest
     ): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = repository.save(
             Note(
                 id = body.id?.let { ObjectId(it) } ?: ObjectId.get(),
@@ -39,7 +41,7 @@ class NoteController(
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId(),
+                ownerId = ObjectId(ownerId),
 
             )
         )
@@ -47,16 +49,22 @@ class NoteController(
         return note.toResponse()
     }
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return repository.findByOwnerId(ObjectId(ownerId)).map{
             it.toResponse()
         }
     }
     @DeleteMapping(path = ["/{id}"])
     fun deleteById(@PathVariable("id") id: String) {
-        repository.deleteById(ObjectId(id))
+        val note = noteRepository.findById(ObjectId(id)).orElseThrow{
+            IllegalArgumentException("Note not found")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if(note.ownerId.toHexString() == ownerId){
+            repository.deleteById(ObjectId(id))
+        }
+
     }
 }
